requirements.txt
requests==2.31.0
openai==1.12.0
python-dotenv==1.0.1
pandas==2.2.1
matplotlib==3.8.3
streamlit==1.32.0